Add confetti and task recommendation dialog features
Introduces a confetti overlay widget triggered on task completion and a dialog for task recommendations based on user energy input. Adds new dependencies: confetti and flutter_rating_bar. Updates navigation from task details to goal view, and refactors task recommendation logic to use double for energy.

+++ lib/src/services/task_recomendation.dart
@@ -3,7 +3,7 @@ import 'package:nava/src/models/task.dart';
 import 'package:nava/src/services/get_tasks_left_in_stage.dart';
 import 'dart:math';
 
-Task? taskRecommendation(int goalId, int energy) {
+Task? taskRecommendation(int goalId, double energy) {
   Stage? currentStage = getCurrentStage(goalId);
   if (currentStage == null) {
     return null;
@@ -65,7 +65,7 @@ double energyCost(Task task){
   return timeCost + difficultyCost;
 }
 
-int obtainAvailableEnergy(int energy){
+int obtainAvailableEnergy(double energy){
   if(energy == 1){
     return 3;
   }else if(energy == 2){
